requirements/base.txt
django==4.2.*
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-environ
django-filter
psycopg2
django-redis
channels
channels-redis
daphne
celery
django-celery-beat
flower
elasticsearch-dsl
django-elasticsearch-dsl
django-storages
boto3
groq
weasyprint
razorpay
django-axes
drf-spectacular
Pillow
python-magic
langsmith
langchain
langchain-core
langchain-groq
langgraph
